% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khanetal2022tethysSSPRCP.R
\docType{package}
\name{khanetal2022tethysSSPRCP}
\alias{khanetal2022tethysSSPRCP}
\title{khanetal2022tethysSSPRCP}
\description{
\itemize{
\item Github: https://github.com/JGCRI/khan-etal_2022_tethysSSPRCP
\item Webpage: https://khan-etal_2022_tethysSSPRCP.github.io/
}
}
